Initial commit of code, as it was left in April 2001
It appears that I left the code in a bit of a state at the point
when I must have found something better to do with myself.
It doesn't compile due to a series of simple syntax errors and
undeclared symbols etc.
I was obviously in the middle of some big alterations. Hurray for
version control, shame I wasn't using it at school.