Add generic file handler class to replace handlers for individual file types
At the moment FileHandler is hard code to read and write puzzle records. Later it will become a
Java collections interface to deal with other record types.
Using the handler the file is now read and written as a serialised object. The same API is
largely maintained in the file handler class .

Add functionality to the ManagePuzzlesDialog and EditPuzzlesDialog to briefly test the file
handling.

Rename the file record classes for clarity.